All these files are necessary to create the process
I think I would have been beneficial to have checked my code in earlier to show the evolution of the development of this process.  As it stands, the code accomplished what it is supposed to do, but feel like I could have been cleaner especially with the functions going through the results of API / json.  I did not do this in an object oriented fashion but at this stage I think that is ok.